Refactor code in createModel + failed initModel use
Change the name of the arguments for the number of cores for more
clarity in the code.
Put parallel initialisation of a block in a function to ease
maintainance.
Update documentation.

NOTE: tried to use initModel instead of having the initialisation code
directly in createModel, but doing so generates a "cannot convert to
environment" Rcpp error.