Almost complete Implementation of the ModelSet fitting
MISSING: ModelSetWrapper::fitmodel has to use a copy of a Data object,
and cannot use a DataSet because Rcpp does not allow function
overloading. This generates an error of "Data and Design inconsistent"
because the data matrix that matches the design in DataSet is not the main
double_matrix.

Define the DataSet class to contain the various Data and be able to use
the Model::eval function in ModelSet::eval
Implement Rcpp wrappers for the ModelSet and the DataSet classes.
Compute a vector for the Data on initialisation and add a constructor to
Data to initialize it.
Fix the ModelSet::eval and update it to use DataSet
Finish the wrappers in R.