Fix ModelSet::predict (but not the discrepancy with fitmodel yet)
Fix ModelSet::predict (there was an error in the shift) by calling the Model::predict.
Add copy_to_view which does not resize, to copy into a view.
Make ModelWrapper::simulate use a Data* instead of Data to be able to
use it with *DataSet.
Replace DataSet* cast by dynamic_cast<DataSet*>