Accept R data to create models + unit test for this feature
Create extraction function to accept both filename and direct R data,
with documentation.
Include MIN_CV and DEFAULT_CV as arguments for the functions
'createModel', 'createModelSet' and 'createModelSet' (TODO add to the model
object as info and use in 'rebuildModel')
Add the possibility to name the model when R object is used.
Change every print to message to fit CRAN good practice.
Get rid of the sampleWithCorrelation function which was not used.
Thorough unit tests for the extraction functions and for the creation
with R objects.
Give the correlated links as info instead of message.
Update 'extractModelCore' documentation to link to the MIDAS description
in 'extractMIDAS'.
Add extra controls to 'extractModelCore' to accept no blank values,
verify that control values exist, accept blank or control without
replicates, and use the extraction functions.
Update 'getModelStructure' and the 'ModelStructure' C++ object to take a
structure name.
Add helper functions 'is.string' and 'checkMIDAS'.
Fix typos and add contexts to the STASNet global test.