Change from chi-2 to T-test and add default/min CV control
Add MIN_CV and DEFAULT_CV as arguments to createModel and
createModelSet.
Add min_cv and default_cv to the MRAmodel, import and export to refit
with the same conditions.
Add control of min and default CV in fit_script.
Move the infos generation to a side function.
Change Chi-2 test to F-test and compute reduced chi-score.
Add tests for the cv fields, the reduced chi-score and the model
reduction.
Update documentation.