Fix the columns names in simulateModel output
The names were not properly attributed, leading to erroneous subsetting.
Put data to the left in plotSimulation.
Add na.rm to min and max in plotSimulation to still be able to plot the
annotations when NAs are present in the data.
Fix when readouts are duplicated and add corresponding unit test.