Fix bug for limited cases for 'no' and 'bystim' reordering
Fix that not everything was properly switched for 'bystim' reordering.
Fix that 'bystim' and 'byinhib' were inverted
Fix that 'no' ordering did not return a matrix for stim_sort and
inh_sort.
Add tests to check this behaviour.
Remove the reordering of the measurements to use the order from the
input file.