Add the possibility to partially refit a model
Add keep_constant in parallel_initialisation.
Change forgotten write to message in parallel_initialisation.
Add collapse to pastecoma and pastetab
'refitModel' to refit the model with a given parameter set
and only refit some parameters.
'fitFromModel' to fit a model using the parameters from another model
with the same structure.
Fix 'computeFitScore' to recompute the likelihood when refit_model is
FALSE.