Add the possibility to plot heatmaps without values
Add a show_values option to plotHeatmap and plotModelAccuracy to control
the display of values in the boxes of the heatmaps, with default TRUE.
Add an option -l in fit_script to plot heatmaps with a limited range for
the color scale which does not plot the values by default.
Fix a typo in fit_script help.
Update documentation.