[incomplete feature] Get the log-fitting to work
Get the log-fitting to work for the MRAmodel (but not ModelSet yet).
Use the standard deviation of the log as the error in log-space,
empirical evidence show that it is very conserved between replicates.
The fitting setting is exported to and imported from the .mra file by
export/import/buildModel.
ModelWrapper have to know about the log-fitting in order for the link
extension subroutine works.

The error from the previous commit was due to a bad recompilation, using
--preclean solved it.