[bugfix] Better fix for importModel spurious warnings
The warnings only occured with log-data because they introduced an NA in
the .mra file, while the linear derivation puts NaN. This is due to the
fact that 'mean(NA)==NaN'. 'geom_mean' has been corrected to reproduce
this behaviour.
Loading .mra files generated with older version will still exhibit the
warning but this feels like a better solution for error management.