[bugfix] Fix unit tests problems linked to log-space fitting
Fix unit test broken by the change of setModel introduced for log-space
fitting.
Fix the error being NA when there is no replicates for log-space fitting.
Fix geom_mean to effectively remove NAs for the computation.
Write a new unit test for createModel with 'data_space="log"'.
Use lower DEFAULT_CV for createModelSet to still be able to recover the
variable link that is expected (but the error model is not known because
it is simulated data).