[feature] rebuildModel using the data from .mra is now possible
Unify residual plots between MRAmodel and MRAmodelSet.
Take the number of replicates into account for log fitting.
Add the possibility to rebuild model without providing an data file if
the data has been saved in the .mra file (which should be the case for
all recently exported models).
Add a test of this new rebuilding option.
Update the documentation