[feature] Improve refitModel behaviour and provide compareParameters generic
Make compareParameters a generic function with implementations for
modelGroup and MRAmodeoSet.
NOTE: the modelGroup implementation makes the assumption that the models
are directly comparable and should be improved.
Improve the residual plot of refitModel.
Change behaviour of refitModel to return an error if no parameters
should be refitted.
Fix some typos is the documentation.
Update documentation.